web-app-example/api/requirements.txt
Flask==2.2.5
redis==4.4.4
gunicorn==20.1.0
flask-cors==3.0.10
psycopg2-binary==2.9.5
requests==2.28.2